01_guardrails/00_nemo_guardrails_exploration/requirements.txt
langchain==0.1.14
langchain-openai==0.0.8
boto3==1.34.64
python-dotenv==1.0.1
pdfplumber==0.10.3
langchainhub
nemoguardrails==0.8.2
chromadb==0.4.24
streamlit==1.32.2
azure-identity==1.15.0
azure-search-documents==11.4.0
boto3==1.34.64